Various changes to remove plaso specifics #1
