+++ .travis.yml
@@ -1,0 +1,13 @@
+language: python
+python:
+  - "2.7"
+before_install:
+  - if test `uname -s` = 'Linux'; then sudo add-apt-repository ppa:gift/dev -y && sudo apt-get update -q && sudo apt-get install libregf-python python-construct python-coverage python-coveralls python-docopt python-six; fi
+script:
+  - ./run_tests.py
+  - ./setup.py build
+  - ./setup.py sdist
+  - ./setup.py bdist
+  - coverage run --source=dfwinreg --omit="*_test*,*__init__*,*test_lib*" ./run_tests.py
+after_success:
+  - coveralls --verbose
