+++ appveyor.yml
@@ -1,0 +1,19 @@
+environment:
+  matrix:
+    - PYTHON: "C:\\Python27"
+
+install:
+  - cmd: '"C:\Program Files\Microsoft SDKs\Windows\v7.1\Bin\SetEnv.cmd" /x86 /release'
+  - ps: (new-object net.webclient).DownloadFile('https://raw.github.com/pypa/pip/master/contrib/get-pip.py', 'C:\Projects\get-pip.py')
+  - ps: (new-object net.webclient).DownloadFile('https://downloads.sourceforge.net/project/pywin32/pywin32/Build%20219/pywin32-219.win32-py2.7.exe', 'C:\Projects\pywin32-219.win32-py2.7.exe')
+  - ps: (new-object net.webclient).DownloadFile('https://pypi.python.org/packages/any/W/WMI/WMI-1.4.9.win32.exe', 'C:\Projects\WMI-1.4.9.win32.exe')
+  - cmd: "%PYTHON%\\python.exe C:\\Projects\\get-pip.py"
+  - cmd: "%PYTHON%\\Scripts\\easy_install.exe C:\\Projects\\pywin32-219.win32-py2.7.exe"
+  - cmd: "%PYTHON%\\Scripts\\easy_install.exe C:\\Projects\\WMI-1.4.9.win32.exe"
+  - cmd: git clone https://github.com/log2timeline/l2tdevtools.git && move l2tdevtools ..\
+  - cmd: mkdir dependencies && set PYTHONPATH=..\l2tdevtools && "%PYTHON%\\python.exe" ..\l2tdevtools\tools\update.py --download-directory dependencies --machine-type x86 --msi-targetdir "%PYTHON%" construct pyregf six
+
+build: off
+
+test_script:
+  - "%PYTHON%\\python.exe run_tests.py"
